refactor: UI 하위에 renderers, handlers 폴더 추가
